refactor(vfs): relocate device headers to public include
