+++ include/rex/filesystem/devices/disc_image_entry.h
