+++ include/rex/filesystem/devices/null_entry.h
