+++ include/rex/filesystem/devices/stfs_container_entry.h
