+++ include/rex/filesystem/devices/stfs_container_file.h
