+++ src/filesystem/devices/host_path_device.cpp
@@ -9,7 +9,7 @@
  * @modified    Tom Clay, 2026 - Adapted for ReXGlue runtime
  */
 
-#include "host_path_entry.h"
+#include <rex/filesystem/devices/host_path_entry.h>
 
 #include <algorithm>
