+++ src/filesystem/devices/stfs_container_device.cpp
@@ -9,8 +9,8 @@
  * @modified    Tom Clay, 2026 - Adapted for ReXGlue runtime
  */
 
-#include "stfs_container_device.h"
-#include "stfs_container_entry.h"
+#include <rex/filesystem/devices/stfs_container_device.h>
+#include <rex/filesystem/devices/stfs_container_entry.h>
 
 #include <algorithm>
 #include <queue>
